requirements.txt
python-telegram-bot==13.2
requests==2.31.0
python_dotenv==1.0.1